libc:  Make __{,u}int32_t long for compatibility
In older versions of Universal Interfaces, CGBase.h defines u_int32_t as
an alias of unsigned long, without checking for a conflicting definition
from <machine/types.h> as it does in the final version.  Such a check is
required in OS X, where u_int32_t aliases unsigned int.  Whereas classic
Mac OS development environments like CodeWarrior define int32_t as long,
not int, and uint32_t as unsigned long.  If u_int32_t is even defined at
all, it aliases uint32_t, and therefore doesn't conflict with CGBase.h's
definition.

Make __uint32_t (and consequently u_int32_t) unsigned long (and likewise
for __int32_t) for compatibility with CGBase.h and Metrowerks' stdint.h.

+++ lamp/POSIX-headers/machine/_types.h
@@ -18,12 +18,12 @@ typedef va_list __va_list;
 
 typedef   signed char       __int8_t;
 typedef   signed short      __int16_t;
-typedef   signed int        __int32_t;
+typedef   signed long       __int32_t;
 typedef   signed long long  __int64_t;
 
 typedef unsigned char       __uint8_t;
 typedef unsigned short      __uint16_t;
-typedef unsigned int        __uint32_t;
+typedef unsigned long       __uint32_t;
 typedef unsigned long long  __uint64_t;
 
 typedef unsigned long long __clock_t;
